INDY-1876: on_catch_up_finished refactoring, update tests
